+++ pharmacy_management/src/main/java/com/backend/pharmacy_management/model/service/manufacturer/IManufacturerService.java
@@ -12,10 +12,10 @@ public interface IManufacturerService {
     Page<Manufacturer> findByManufacturerCode(String code,Pageable pageable );
     Page<Manufacturer> findByManufacturerAddress( String address,Pageable pageable);
     Page<Manufacturer> findByManufacturerPhoneNumber( String phoneNumber,Pageable pageable);
-    Manufacturer findByManufacturerId( Integer id);
+    Manufacturer findByManufacturerId( Long id);
     Page<Manufacturer> findByManufacturerNote( String note,Pageable pageable);
     Manufacturer saveManufacturer(Manufacturer manufacturer);
-    void deleteManufacturer(Integer id);
+    void deleteManufacturer(Long id);
 
     Page<Manufacturer> findAllSortName(Pageable pageable);
 
@@ -26,7 +26,7 @@ public interface IManufacturerService {
     Page<Manufacturer> findAllSortNote(Pageable pageable);
 
     Page<Manufacturer> findAllSortPhoneNumber(Pageable pageable);
-    Page<ImportBill> findByIdManufacturer(Integer id,Pageable pageable);
-    ImportBill findByIdImportBill(Integer id);
-    Page<ImportBill> findByDateImportBill(Integer id,String startDate, String endDate,Pageable pageable);
+    Page<ImportBill> findByIdManufacturer(Long id,Pageable pageable);
+    ImportBill findByIdImportBill(Long id);
+    Page<ImportBill> findByDateImportBill(Long id,String startDate, String endDate,Pageable pageable);
 }
